requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
mysql-connector-python
